feat: change ts structure
