+++ src/index.html
@@ -7,38 +7,5 @@
   <title>Online-store</title>
 </head>
 <body id="body">
-
-  <div class="sort-menu">
-
-    <div class="select-box">
-      <div class="options-container">
-
-        <div class="option">
-          <input type="radio" class="radio" id="priseLTH" name="category" />
-          <label for="priseLTH">Prise (low to high)</label>
-        </div>
-
-        <div class="option">
-          <input type="radio" class="radio" id="priseHTL" name="category" />
-          <label for="priseHTL">Prise (high to low)</label>
-        </div>
-
-        <div class="option">
-          <input type="radio" class="radio" id="Rating" name="category" />
-          <label for="Rating">Rating</label>
-        </div>
-
-        <div class="option">
-          <input type="radio" class="radio" id="release" name="category" />
-          <label for="release">Release date</label>
-        </div>
-      </div>
-
-      <div class="selected">
-        Sort by
-      </div>
-    </div>
-  </div>
-  
 </body>
 </html>
